+++ src/main/java/duke/command/ExitCommand.java
@@ -1,0 +1,14 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class ExitCommand extends Command {
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        storage.save(taskList);
+        ui.goodBye();
+        System.exit(0);
+    }
+}
