+++ src/main/java/duke/command/ListCommand.java
@@ -1,0 +1,13 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class ListCommand extends Command {
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        taskList.printList();
+    }
+}
